# Optimize decomposition calculations to reduce model train time by ~23% (Python version)