Summary
This PR optimizes _calculate_decomp_spend_dist and _calculate_x_decomp_agg, reducing redundant computations and improving overall model execution time by ~23% on python version.

### Fixes
This PR does not address a specific GitHub issue but significantly enhances efficiency.

### Type of change
fix: Performance optimization (non-breaking change that improves efficiency)

### Motivation & Context
Decomposition calculations were a major performance bottleneck, slowing down overall execution.

### Key optimization:
Precomputed denominator for xDecompMeanNon0Perc, eliminating redundant computations inside loops.

### Performance Benchmark (tutorial1.ipynb, 5 trials, 2000 iterations per trial)
Before Optimization:

_calculate_decomp_spend_dist: 116.5s
_calculate_x_decomp_agg: 318.1s
_run_nevergrad_optimization: 1304s
After Optimization:

_calculate_decomp_spend_dist: 72.43s (~38% faster)
_calculate_x_decomp_agg: 67.4s (~78% faster)
_run_nevergrad_optimization: 1024s (~23% faster)
This results in an overall model train time improvement of ~23%.